Refact functions to chunk.go and metadata.go
