+++ internal/service/file.go
@@ -662,7 +662,7 @@ func (s *FileService) deleteDocumentFromEngine(ctx context.Context, doc *entity.
 	reqCtx, cancel := context.WithTimeout(ctx, 300*time.Second)
 	defer cancel()
 	condition := map[string]interface{}{"doc_id": doc.ID}
-	if _, err := docEngine.Delete(reqCtx, condition, indexName, doc.KbID); err != nil {
+	if _, err := docEngine.DeleteChunks(reqCtx, condition, indexName, doc.KbID); err != nil {
 		return fmt.Errorf("delete document from engine: %w", err)
 	}
 	return nil
