+++ internal/service/skill_indexer.go
@@ -171,7 +171,7 @@ func (s *SkillIndexerService) IndexSkill(ctx context.Context, tenantID, spaceID 
 
 	// For Infinity: ensure table exists with correct dimension BEFORE inserting
 	if docEngine.GetType() == "infinity" {
-		exists, _ := docEngine.StoreExists(ctx, indexName)
+		exists, _ := docEngine.ChunkStoreExists(ctx, indexName, "skill")
 		if !exists {
 			common.Info(fmt.Sprintf("Creating Infinity table with dimension %d", dimension))
 			if err := s.createIndexWithDimension(ctx, tenantID, spaceID, docEngine, embdID, dimension); err != nil {
@@ -252,7 +252,7 @@ func (s *SkillIndexerService) BatchIndexSkills(ctx context.Context, tenantID, sp
 	if docEngine.GetType() == "infinity" {
 		// For Infinity: must ensure table exists with correct dimension BEFORE inserting
 		common.Info(fmt.Sprintf("Checking if index exists: %s", indexName))
-		exists, err := docEngine.StoreExists(ctx, indexName)
+		exists, err := docEngine.ChunkStoreExists(ctx, indexName, "skill")
 		if err != nil {
 			common.Warn(fmt.Sprintf("Error checking index existence: %v", err))
 		}
@@ -436,10 +436,10 @@ func (s *SkillIndexerService) ReindexAll(ctx context.Context, tenantID, spaceID 
 
 	// Delete existing index and recreate with new dimension (for both ES and Infinity)
 	indexName := getSkillIndexName(tenantID, spaceID)
-	exists, _ := docEngine.StoreExists(ctx, indexName)
+	exists, _ := docEngine.ChunkStoreExists(ctx, indexName, "skill")
 	if exists {
 		common.Info(fmt.Sprintf("ReindexAll: deleting existing index %s", indexName))
-		if err := docEngine.DropStore(ctx, indexName); err != nil {
+		if err := docEngine.DropChunkStore(ctx, indexName, "skill"); err != nil {
 			common.Warn(fmt.Sprintf("ReindexAll: failed to delete existing index: %v", err))
 		}
 	}
@@ -846,7 +846,7 @@ func (s *SkillIndexerService) InitializeIndex(ctx context.Context, tenantID, spa
 
 	common.Info("Checking skill index existence", zap.String("indexName", indexName), zap.String("tenantID", tenantID), zap.String("spaceID", spaceID))
 
-	exists, err := docEngine.StoreExists(ctx, indexName)
+	exists, err := docEngine.ChunkStoreExists(ctx, indexName, "skill")
 	if err != nil {
 		common.Error("Failed to check index existence", err)
 		return fmt.Errorf("failed to check index existence: %w", err)
@@ -883,14 +883,14 @@ func (s *SkillIndexerService) createIndexWithDimension(ctx context.Context, tena
 
 	// For Infinity: check if table exists and needs recreation (dimension mismatch)
 	if docEngine.GetType() == "infinity" {
-		exists, err := docEngine.StoreExists(ctx, indexName)
+		exists, err := docEngine.ChunkStoreExists(ctx, indexName, "skill")
 		if err != nil {
 			common.Warn(fmt.Sprintf("Error checking if index exists: %v", err))
 		}
 		if exists {
 			common.Info(fmt.Sprintf("Index exists, deleting for recreation with dimension %d", dimension),
 				zap.String("indexName", indexName))
-			if err := docEngine.DropStore(ctx, indexName); err != nil {
+			if err := docEngine.DropChunkStore(ctx, indexName, "skill"); err != nil {
 				common.Warn(fmt.Sprintf("Failed to delete existing index: %v", err))
 			}
 		}
