+++ internal/service/skill_space.go
@@ -523,7 +523,7 @@ func (s *SkillSpaceService) asyncDeleteSpace(spaceID, folderID, tenantID string,
 		indexName := getSkillIndexName(tenantID, spaceID)
 		common.Info("Async deleting space index", zap.String("index", indexName), zap.String("spaceID", spaceID))
 		deleteCtx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-		if err := docEngine.DropStore(deleteCtx, indexName); err != nil {
+		if err := docEngine.DropChunkStore(deleteCtx, indexName, "skill"); err != nil {
 			common.Warn("Failed to delete space index during async delete", zap.String("index", indexName), zap.Error(err))
 			// Continue with other cleanup steps
 		} else {
